Inkarnation/Server/requirements.txt
dotenv
numpy
pydantic
pillow
aiohttp